finished adding a place
