+++ src/main/java/com/google/codeu/servlets/PlaceServlet.java
@@ -8,23 +8,58 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.List;
 import java.util.Properties;
 import com.google.appengine.api.users.UserService;
 import com.google.appengine.api.users.UserServiceFactory;
 import com.google.codeu.data.Datastore;
 import com.google.codeu.data.Message;
+import com.google.codeu.data.User;
+import com.google.gson.Gson;
 import org.jsoup.Jsoup;
 import org.jsoup.safety.Whitelist;
 
 import com.google.codeu.data.Place;
 
 @WebServlet("/api/place")
-public class AddPlaceServlet extends HttpServlet {
+public class PlaceServlet extends HttpServlet {
     private Datastore datastore;
 
     @Override
     public void init() {
         datastore = new Datastore();
+    }
+
+
+    /**
+     * Responds with a JSON representation of {@link Message} data for a specific
+     * user. Responds with an empty array if the user is not provided.
+     */
+    @Override
+    public void doGet(HttpServletRequest request, HttpServletResponse response)
+            throws IOException {
+
+        response.setContentType("application/json");
+
+        String email = request.getParameter("user");
+        List<Place> places;
+        if (email == null || email.isEmpty()) {
+            // Request is invalid, return empty array
+            // esponse.getWriter().println("[]");
+            places = Place.getAll();
+        } else {
+            User user = User.getByEmail(email);
+            if(user == null){
+                response.setStatus(400);
+                return;
+            }
+            places = Place.getByUser(User.getByEmail(email));
+        }
+
+        Gson gson = new Gson();
+        String json = gson.toJson(places);
+
+        response.getWriter().println(json);
     }
 
     @Override
